% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem-utils.R
\name{dem_as_image}
\alias{dem_as_image}
\title{dem_as_image}
\usage{
dem_as_image(aoi, dem.src = c("glo30", "srtm"), mask = FALSE)
}
\arguments{
\item{dem.src}{}
}
\value{

}
\description{
create a DEM image - choose from different sources. Currently only GLO30
and SRTM.
}
