% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{ee_ahist}
\alias{ee_ahist}
\title{ee_ahist}
\usage{
ee_ahist(img, ee_fc = NULL)
}
\arguments{
\item{img}{}
}
\value{
a ggplot object with the histogram(s)
}
\description{
gives approximate histograms for all bands in an ee image based on random points
}
