% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{est.err}
\alias{est.err}
\title{estimation error}
\usage{
est.err(C, Chat)
}
\arguments{
\item{C}{true coefficient}

\item{Chat}{estimated coefficient}
}
\value{
estimation error
}
\description{
estimation error
}
