% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-utils.R
\name{focalSpeckle}
\alias{focalSpeckle}
\title{focalSpeckle}
\usage{
focalSpeckle(
  img,
  SMOOTHING_RADIUS = 35,
  bands = c("VV", "VH"),
  focal.type = c("mean", "median", "mode", "min", "max")
)
}
\arguments{
\item{img}{and ee image}

\item{SMOOTHING_RADIUS}{default 35.}

\item{bands}{The bands over which to undertake smoothing.}

\item{focal.type}{default 'mean'. The function to use over the focal area.}
}
\value{

}
\description{
speckle filter for SAR imagery.
}
