% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{frate}
\alias{frate}
\title{FDR and FNR}
\usage{
frate(U, Uhat, dhat)
}
\arguments{
\item{U}{left singular vectors}

\item{Uhat}{estimated left singular vectors}

\item{dhat}{estimated singualr values}
}
\value{
\item{rate}{consisting FDR and FNR}
}
\description{
FDR and FNR
}
