% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-correct-terrain.R
\name{illuminationCondition}
\alias{illuminationCondition}
\title{illuminationCondition}
\usage{
illuminationCondition(img, dem, buf = 10000)
}
\arguments{
\item{buf}{}
}
\value{
...
}
\description{
illuminationCondition
}
