% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-correct-terrain.R
\name{illuminationCorrection}
\alias{illuminationCorrection}
\title{Correct Landsat 8}
\usage{
illuminationCorrection(img, dem, buf = 10000)
}
\arguments{
\item{buf}{}
}
\value{

}
\description{
Function to apply the Sun-Canopy-Sensor + C (SCSc) correction method to a
landsat 8 image. Function by Patrick Burns and Matt Macander.
}
