% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{img_band_names}
\alias{img_band_names}
\title{img_band_names}
\usage{
img_band_names(img)
}
\arguments{
\item{img}{}
}
\value{
character vector of band names.
}
\description{
get ee_image band names as a raster
}
