% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuamut}
\alias{kuamut}
\title{Kuamut Conservation Area}
\format{
An sf object of the AOI
}
}
\source{
permian global internal
}
\usage{
kuamut
}
\description{
A spatial vector as an sf object denoting a conservation region in Malaysia
}
\keyword{datasets}
