% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-utils.R
\name{l8_collect_mask_clouds}
\alias{l8_collect_mask_clouds}
\title{l8_collect_mask_clouds}
\usage{
l8_collect_mask_clouds(img.coll)
}
\arguments{
\item{img.coll}{}
}
\value{

}
\description{
l8_collect_mask_clouds
}
