% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-mask-clouds.R
\name{l8_mask_clouds}
\alias{l8_mask_clouds}
\alias{l8_mask_clouds.ee.image.Image}
\alias{l8_mask_clouds.ee.imagecollection.ImageCollection}
\title{l8_mask_clouds: cloud mask an l8 image or collection}
\usage{
l8_mask_clouds(x, ...)

\method{l8_mask_clouds}{ee.image.Image}(x, ...)

\method{l8_mask_clouds}{ee.imagecollection.ImageCollection}(x, ...)
}
\arguments{
\item{x}{An ee Image or ImageCollection
@param ... Not currently used.}
}
\description{
get projection of spatial object
}
\details{
cloud mask an l8 image or collection
}
