% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-correct-terrain.R
\name{l8_terrain_correct_method}
\alias{l8_terrain_correct_method}
\title{l8_terrain_correct}
\usage{
l8_terrain_correct_method(x, aoi.sf, buf = 10000, mask = FALSE)
}
\arguments{
\item{x}{either an earth engine Collection or Image}

\item{aoi.sf}{an sf or sfc object defining the aoi.}
}
\value{

}
\description{
correct illumination effect for Landsat8 collection.
}
