% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-correct-terrain.R
\name{l8_terrain_dtm}
\alias{l8_terrain_dtm}
\title{l8_terrain_dtm}
\usage{
l8_terrain_dtm(x, aoi.sf)
}
\arguments{
\item{x}{either an earth engine Collection or Image}

\item{aoi.sf}{an sf or sfc object defining the aoi.}
}
\value{
A ee image of SRTM DEM.
}
\description{
correct illumination effect for Landsat8 collection.
}
