% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-mapping.R
\name{MapAll}
\alias{MapAll}
\title{MapAll}
\usage{
MapAll(img, zoom = 11, pal = viridisLite::plasma(256), opacity = 1)
}
\arguments{
\item{img}{an ee image}

\item{pal}{default is \code{viridisLite::plasma(256)}. HEX code colours...}
}
\value{

}
\description{
Map all the bands of an ee image
}
