% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-utils.R
\name{mask_clouds_l8_method}
\alias{mask_clouds_l8_method}
\title{mask_clouds_l8_method}
\usage{
mask_clouds_l8_method(img)
}
\arguments{
\item{img}{an earth engine ee image}
}
\value{
a cloud masked ee image
}
\description{
mask_clouds_l8_method
}
