% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-adugna-speckle.R
\name{MonoTemporal_Filter}
\alias{MonoTemporal_Filter}
\title{Adugna MonoTemporal_Filter}
\usage{
MonoTemporal_Filter(coll, KERNEL_SIZE = 9, SPECKLE_FILTER = "GAMMA MAP")
}
\arguments{
\item{SPECKLE_FILTER}{}
}
\value{

}
\description{
Adugna MonoTemporal_Filter
}
