% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer}
\alias{peer}
\title{Parallel estimation with extracted layers and regularization}
\usage{
peer(X, Y, nrank = 3, penalty = "l1", eps = 1e-06)
}
\arguments{
\item{X}{predictors matrix}

\item{Y}{responses matrix}

\item{nrank}{rank of estimator}

\item{penalty}{penalty type (l1, l0, default l1)}

\item{eps}{tolerance parameter}
}
\value{
\item{C}{estimator for coefficient}
 \item{U}{estimator for left singular vectors}
 \item{V}{estimator for right singular vectors}
 \item{d}{estimator for singular values}
 \item{time}{elapsedtime in parallel version}
}
\description{
Parallel estimation with extracted layers and regularization
}
