% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_init.R
\name{peer_init}
\alias{peer_init}
\title{Initialize ee module to use {peer}}
\usage{
peer_init(user = NULL, ..., gcs = TRUE)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
A wrapper for rgee::ee_Initialize(). Loads rgee and initializes the ee
python module.
}
