% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{point_extract}
\alias{point_extract}
\title{point_extract}
\usage{
point_extract(img, ee_fc)
}
\arguments{
\item{img}{ee image}

\item{ee_fc}{an ee feature collection (points)}
}
\value{
a dataframe of values for all properties
}
\description{
point_extract
}
