% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{pred.err}
\alias{pred.err}
\title{prediction error}
\usage{
pred.err(X, C, Chat)
}
\arguments{
\item{X}{predictors matrix}

\item{C}{true coeffiecient}

\item{Chat}{estimated coefficient}
}
\value{
prediction error
}
\description{
prediction error
}
