% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{rando_points}
\alias{rando_points}
\title{rando_points}
\usage{
rando_points(img, n = 1000)
}
\arguments{
\item{img}{and ee image}

\item{n}{default is 1000. the number of points to generate. Note Earth Engine
has some limits on how many values can be returned.}
}
\value{

}
\description{
creaets random points across the extent of an ee image.
}
