% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{return_ee_geom}
\alias{return_ee_geom}
\title{helper to return ee_featurecollection regardless of class}
\usage{
return_ee_geom(x)
}
\arguments{
\item{x}{either sf or ee.featurecollection.FeatureCollection}
}
\value{
ee.featurecollection.FeatureCollection
}
\description{
helper to return ee_featurecollection regardless of class
}
