% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-adugna-process.R
\name{s1_adugna_process}
\alias{s1_adugna_process}
\alias{s1_adugna_process.ee.imagecollection.ImageCollection}
\title{Create Analysis ready Sentinel 1 image or collection}
\usage{
s1_adugna_process(x, aoi.sf, rtc = TRUE, reduce = TRUE, mask = FALSE, ...)

\method{s1_adugna_process}{ee.imagecollection.ImageCollection}(x, aoi.sf, reduce = TRUE)
}
\arguments{
\item{x}{an ee image or collection}

\item{aoi.sf}{sf object to set extent for clipping.}

\item{rtc}{Logical - should Terrain Correction/Flattening be applied?}

\item{reduce}{Only for image colections - should the collection be reduced to
an image.}

\item{mask}{Not used}

\item{...}{not used}
}
\value{

}
\description{
Using tools from Adugna, et al 2021 (https://www.mdpi.com/2072-4292/13/10/1954)
}
