% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-adugna-process.R
\name{s1_adugna_wrapper}
\alias{s1_adugna_wrapper}
\title{s1 processing wrapper}
\usage{
s1_adugna_wrapper(x)
}
\arguments{
\item{x}{ee image}
}
\value{
ee image
}
\description{
s1 processing wrapper
}
