% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-utils.R
\name{s1_collect}
\alias{s1_collect}
\title{s1_collect}
\usage{
s1_collect(
  aoi,
  start.date,
  end.date,
  orbit.pass = c("DESCENDING", "ASCENDING"),
  dataset = c("S1_GRD", "S1_GRD_FLOAT")
)
}
\arguments{
\item{orbit.pass}{Default 'DESCENDING'. Must be either}
}
\value{

}
\description{
s1_collect
}
