% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-correct-terrain.R
\name{s1_wrapper_rtc}
\alias{s1_wrapper_rtc}
\title{s1 RTC processing wrapper}
\usage{
s1_wrapper_rtc(x)
}
\arguments{
\item{x}{ee image}
}
\value{
ee image
}
\description{
s1 RTC processing wrapper
}
