% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_utils.R
\name{s2_collect}
\alias{s2_collect}
\title{s2_collect}
\usage{
s2_collect(
  aoi,
  start.date,
  end.date,
  min.cloud = 90,
  dataset = c("S2_SR", "S2_SR_HARMONIZED")
)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
s2_collect
}
