% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gcs.R
\name{save_to_gcs}
\alias{save_to_gcs}
\title{Save an ee image to Google Cloud Stoarge}
\usage{
save_to_gcs(
  x,
  filename,
  scale,
  region,
  description,
  bucket = "rgee_store",
  max.wait = 30,
  task_time = 10,
  public = TRUE,
  add.vsi = TRUE
)
}
\arguments{
\item{x}{an ee image}

\item{filename}{filename without extension for GCS}

\item{task_time}{}
}
\value{

}
\description{
Save an ee image to Google Cloud Stoarge
}
