% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gcs.R
\name{setup_gcs}
\alias{setup_gcs}
\title{Set up GOogle Cloud Storage}
\usage{
setup_gcs(sak_file, user, check_bucket = TRUE, bucket_name = "rgee_store")
}
\arguments{
\item{bucket_name}{}
}
\value{

}
\description{
Set up GOogle Cloud Storage
}
