% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{sf_ext_as_ee}
\alias{sf_ext_as_ee}
\title{sf_ext_as_ee}
\usage{
sf_ext_as_ee(x)
}
\arguments{
\item{x}{sf of sfc object.}
}
\value{

}
\description{
convert an sf object into a ee Feature Collection describing the bounding box
of the sf object.
}
