% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.control}
\alias{sim.control}
\title{generate a list consisting simulation setup}
\usage{
sim.control(n, p, q, nrank, s, snr, erank, miss = 0)
}
\arguments{
\item{n}{number of observations}

\item{p}{dimension of predictor vector}

\item{q}{dimension of reponse vector}

\item{nrank}{rank of true coefficient}

\item{s}{sparsity of left singular vectors}

\item{snr}{singnal to noise rate}

\item{erank}{rank for estimator}
}
\value{
\item{control}{a list consisting simulation setup}
}
\description{
generate a list consisting simulation setup
}
