% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.miss}
\alias{sim.miss}
\title{generate table for missing data simulation. Then tbale includes ErC, ErXC, FPR, FNR and time with 
l1 and l0 penalty.}
\usage{
sim.miss(control = sim.control())
}
\arguments{
\item{control}{a list consists simlation setting, which can be generated by sim.control() function}
}
\value{
\item{table}{including ErC, ErXC, FPR, FNR and time with l1 and l0 penalty.}
}
\description{
generate table for missing data simulation. Then tbale includes ErC, ErXC, FPR, FNR and time with 
l1 and l0 penalty.
}
