% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{sim.setup}
\alias{sim.setup}
\title{generate data and coefficient}
\usage{
sim.setup(n = 100, p = 100, q = 100, nrank = 3, s = 4,
  snr = 0.5, miss = 0)
}
\arguments{
\item{n}{number of observations}

\item{p}{dimension of predictor vector}

\item{q}{dimension of reponse vector}

\item{nrank}{rank of true coefficient}

\item{s}{sparsity of left singular vectors}

\item{snr}{singnal to noise rate}

\item{miss}{missing rate of response matrix}

\item{erank}{rank for estimator}
}
\value{
\item{X}{predictor matrix}
 \item{Y}{response matrix}
 \item{d}{singular values}
 \item{U}{left singular vectors}
 \item{V}{right singular vectors}
 \item{C}{coefficient matrix}
}
\description{
generate data and coefficient
}
