% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.table}
\alias{sim.table}
\title{generate result table with RRR, SRRR, SEED, PEER(l1) and PEER(l0), which
includes ErC, ErXC, FPR, FNR and time.}
\usage{
sim.table(control = control)
}
\arguments{
\item{control}{a list consists simlation setting, which can be generated by sim.control() function}
}
\value{
\item{table}{result table with RRR, SRRR, SEED, PEER(l1) and PEER(l0), which includes ErC, ErXC, FPR, FNR and time.}
}
\description{
generate result table with RRR, SRRR, SEED, PEER(l1) and PEER(l0), which
includes ErC, ErXC, FPR, FNR and time.
}
