% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.time}
\alias{sim.time}
\title{conuting time for each methos (SRRR, SEED, PEER(l1), PEER(l0))}
\usage{
sim.time(X, Y, erank)
}
\arguments{
\item{X}{predictors matrix}

\item{Y}{responses matrix}

\item{erank}{rank of estimator}
}
\value{
{Time}{times of SRRR, SEED, PEER(l1), PEER(l0), respectively}
}
\description{
conuting time for each methos (SRRR, SEED, PEER(l1), PEER(l0))
}
