% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-adugna-terrain-flat.R
\name{slope_correction}
\alias{slope_correction}
\title{Adunga S1 slope correction}
\usage{
slope_correction(
  collection,
  TERRAIN_FLATTENING_MODEL = "VOLUME",
  DEM = ee$Image("USGS/SRTMGL1_003"),
  TERRAIN_FLATTENING_ADDITIONAL_LAYOVER_SHADOW_BUFFER = 0
)
}
\arguments{
\item{TERRAIN_FLATTENING_ADDITIONAL_LAYOVER_SHADOW_BUFFER}{}
}
\value{

}
\description{
Adunga S1 slope correction
}
