% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{sub_bounds_func}
\alias{sub_bounds_func}
\title{clip an image add properties.}
\usage{
sub_bounds_func(aoi)
}
\arguments{
\item{aoi}{}
}
\value{
clipped img
}
\description{
clip an image add properties.
}
