% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{subset_bounds}
\alias{subset_bounds}
\title{clip an image add properties.}
\usage{
subset_bounds(img, aoi)
}
\arguments{
\item{aoi}{}
}
\value{
clipped img
}
\description{
clip an image add properties.
}
