% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-correct-terrain.R
\name{terrain_flatten}
\alias{terrain_flatten}
\title{Terrain flatenning for s1 image}
\usage{
terrain_flatten(img)
}
\arguments{
\item{img}{an ee image}
}
\value{
a terrain flattened image with additional "layover" mask band.
}
\description{
Applies terrain flattening alg by Andreas Vollrath (ESA) to an ee image.
}
\details{
Implementation by Andreas Vollrath (ESA), inspired by Johannes Reiche (Wageningen)
}
