% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-helpers.R
\name{ee_ahist}
\alias{ee_ahist}
\title{ee_ahist}
\usage{
ee_ahist(img, ee_fc = NULL, n = 1000)
}
\arguments{
\item{img}{an earth engine image to sample from}

\item{ee_fc}{an earth engine feature class of points to use. Useful when
comparing between images to maintain consistent sampling. Whn NULL, the points
are randomly generated.}

\item{n}{The number of random points to sample if ee_fc is NULL. Lower this value
when the this error occurs: "User memory limit exceeded."}
}
\value{
a ggplot object with the histogram(s)
}
\description{
gives approximate histograms for all bands in an ee image based on random points
}
