% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l8-correct-terrain.R
\name{l8_terrain_correct}
\alias{l8_terrain_correct}
\title{l8_terrain_correct}
\usage{
l8_terrain_correct(
  x,
  aoi.sf,
  dem.src = c("glo30", "srtm"),
  buf = 10000,
  mask = FALSE
)
}
\arguments{
\item{x}{either an earth engine Collection or Image}

\item{dem.src}{}
}
\value{

}
\description{
correct illumination effect for Landsat8 collection.
}
