% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-correct-terrain.R
\name{s1_process}
\alias{s1_process}
\alias{s1_process.ee.image.Image}
\alias{s1_process.ee.imagecollection.ImageCollection}
\title{Process Sentinel 1 image or collection}
\usage{
s1_process(x, aoi.sf, mask = FALSE, ...)

\method{s1_process}{ee.image.Image}(x, aoi.sf)

\method{s1_process}{ee.imagecollection.ImageCollection}(x, aoi.sf)
}
\arguments{
\item{x}{an ee image or collection}

\item{aoi.sf}{an sf/sfc object of the area of interest}

\item{...}{}
}
\value{

}
\description{
Process Sentinel 1 image or collection
}
