% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1-correct-terrain.R
\name{s1_process}
\alias{s1_process}
\alias{s1_process.ee.image.Image}
\alias{s1_process.ee.imagecollection.ImageCollection}
\title{Process Sentinel 1 image or collection}
\usage{
s1_process(x, reduce = TRUE, mask = FALSE, ...)

\method{s1_process}{ee.image.Image}(x)

\method{s1_process}{ee.imagecollection.ImageCollection}(x, reduce = TRUE)
}
\arguments{
\item{x}{an ee image or collection}

\item{reduce}{Only for image colections - should the collection be reduced to
an image.}

\item{mask}{Not used}

\item{...}{not used}
}
\value{

}
\description{
Process Sentinel 1 image or collection
}
